requirements.txt
Django>=3.2.9,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9.2,<2.9.5
Pillow>=8.4.0,<8.4.5

flake8>=4.0.0,<4.1.0